realease 0.5.1 working on conda release
