+++ setup.py
@@ -21,7 +21,7 @@ def find_files(directory):
 
 setup(
     name="meeko",
-    version='0.5.0',
+    version='0.5.1',
     author="Forli Lab",
     author_email="forli@scripps.edu",
     url="https://github.com/ccsb-scripps/meeko",
@@ -29,16 +29,6 @@ setup(
     long_description=open(os.path.join(base_dir, 'README.md')).read(),
     long_description_content_type="text/markdown",
     packages=find_packages(),
-    scripts=["scripts/mk_prepare_ligand.py",
-             "scripts/mk_prepare_receptor.py",
-             "scripts/mk_export.py",
-           #"scripts/dry.py",
-           #"scripts/mapwater.py",
-           #"scripts/wet.py"],
-            ],
-    package_data={"meeko" : ["data/*"]},
-    data_files=[("", ["README.md", "LICENSE"]),
-                ("scripts", find_files("scripts"))],
     include_package_data=True,
     zip_safe=False,
     install_requires=['numpy>=1.18'],
@@ -66,5 +56,12 @@ setup(
         'Topic :: Scientific/Engineering :: Bio-Informatics',
         'Topic :: Scientific/Engineering :: Chemistry',
         'Topic :: Software Development :: Libraries'
-    ]
+    ],
+    entry_points={
+        'console_scripts': [
+            'mk_export=meeko.cli.mk_export:main',
+            'mk_prepare_ligand=meeko.cli.mk_prepare_ligand:main',
+            'mk_prepare_receptor=meeko.cli.mk_prepare_receptor:main'
+        ]
+    }
 )
